requirements.txt
streamlit>=1.22.0
python-dotenv>=0.21.0
requests>=2.28.1